host-ct-ng: turn off DEVPKG
A package with DEVPKG set gets a configured with "--prefix=" so that it
can be installed with DESTDIR into its sysroot. The idea is to make the
pacakge relocateable. It may then be packed and in a later run reused,
i.e. unpacked.

However the hand-crafted configure of ct-ng doesn't cope with "--prefix=".
So disable the DEVPKG feature.

Reported-by: Jon Ringle <jon@ringle.org>
Signed-off-by: Marc Kleine-Budde <mkl@pengutronix.de>
Signed-off-by: Jon Ringle <jon@ringle.org>